requirements.txt
Flask==1.1.1
Flask-RESTful==0.3.8
gunicorn==19.9.0
joblib==0.14.1
json5==0.9.1
jsonschema==3.2.0
Keras==2.3.1
Keras-Applications==1.0.8
Keras-Preprocessing==1.1.0
keyring==21.1.0
lxml==4.5.0
numpy==1.18.1
pandas==1.0.1
path==13.1.0
pickleshare==0.7.5
python-dateutil==2.8.1
python-jsonrpc-server==0.3.4
python-language-server==0.31.7
requests==2.22.0
requests-oauthlib==1.3.0
scikit-image==0.16.2
scikit-learn==0.22.1
scipy==1.3.1
tensorboard==2.0.2
tensorboard-plugin-wit==1.6.0.post3
tensorflow==2.0.0
tensorflow-estimator==2.0.1
azure-storage-blob==2.1.0